+++ Consola1/Sumar.cs
@@ -13,9 +13,8 @@ namespace Consola1
             darSalida(Resultado);
         }
 
-        static int operacion(int a, int b)
+        static float operacion(float a, float b)
         {
-
             return a+ b;
         }
     }
